study buddy/falcon-tdova-nov25-livekit/backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.1
PyJWT==2.8.0
requests==2.31.0
